Raihan/auth

+++ src/main/java/id/ac/ui/cs/advprog/authjwt/service/AuthService.java
@@ -22,9 +22,6 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.HashMap;
-import java.util.Map;
-
 @Service
 public class AuthService implements AuthenticationFacade {
     @Autowired
@@ -77,6 +74,11 @@ public class AuthService implements AuthenticationFacade {
     @Override
     public ResponseEntity<LogoutResponseDTO> logout(LogoutRequestDTO token){
         try{
+
+            if(tokenRepository.findByToken(token.token()) == null) {
+                throw new Exception("Token not found");
+            }
+
             tokenRepository.deleteByToken(token.token());
             return new ResponseEntity<>(
                     new LogoutResponseDTO("accept", "Succes to logout"),
